pyproject.toml
[project]
name = "automaton-auditor-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython",
    "httpx",
    "langchain",
    "langchain-google-genai",
    "langchain-openai",
    "langgraph",
    "pydantic",
    "python-dotenv",
    "pyyaml",
    "rich",
    "pymupdf",
    "tree-sitter",
    "tree-sitter-python",
    "google-genai>=1.64.0",
    "langsmith>=0.7.6",
    "langchain-groq>=1.1.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
